feat:add diary 기능에 session 권한 적용

+++ app/route/diary_route.py
@@ -17,9 +17,10 @@ def generate_dates(year, month):
 # 일기 페이지 라우트
 @diary_bp.route('/')
 def diary_home():
-    # if "user_id" not in session:
-    #     return redirect(url_for())
-    # else:
+    if "user_id" not in session:
+        return redirect(url_for('auth.login_user'))
+    else:
+        user_id = session['user_id']
         year = request.args.get('year', datetime.now().year, type=int)
         month = request.args.get('month', datetime.now().month, type=int)
 
@@ -35,54 +36,42 @@ def diary_home():
         # 해당 년월의 날짜 리스트 생성 (정수형 일(day)만 반환)
         days_in_month = generate_dates(year, month)
 
-        # 테스트
-        user_id = 1
-
         diaries = DiaryDAO().get_list_diaries_with_date(user_id, year, month)       
         return render_template('diary.html', year=year, month=month, days_in_month=days_in_month, diaries=diaries)
 
 # 특정 날짜에 대한 일기 작성 페이지 라우트
 @diary_bp.route('/write/<int:day>', methods=["GET", "POST"])
 def write_diary(day):
-    # if "user_id" not in session:
-    #     # return redirect(url_for())
-    #     return
-    # else:
+    if 'user_id' not in session:
+        return redirect(url_for('auth.login_user'))
+    else:
         year = request.args.get('year', datetime.now().year, type=int)
         month = request.args.get('month', datetime.now().month, type=int)
 
         diary_id = request.args.get('diary_id', None)
-        _method = request.form.get('_method', 'ㅇㅇ')
-        print(_method)
+        user_id = session['user_id']
+        _method = request.form.get('_method', 'upsert')
 
-
-        # 권한 추가
+        if diary_id:
+            diary = DiaryDAO().get_diary(diary_id)
+            if diary['user_id'] != user_id:
+                abort(403)
+        else:
+            diary = None
+                        
         if request.method == 'GET':
-            # formatted_date = datetime(year, month, int(day)).strftime('%Y-%m-%d')
-
-            if diary_id:
-                 diary = DiaryDAO().get_diary(diary_id)
-            else:
-                 diary = None
-
             return render_template('write_diary.html', year=year, month=month, day=day, diary=diary)
-
         elif _method == "upsert":
-            user_id = 1
-            # user_id = session['user_id']
             mood = request.form['mood']
             body = request.form['body']
             # file_urls = request.form['file_urls']
-            formatted_date = datetime(year, month, int(day)).strftime('%Y-%m-%d')
-            
-            DiaryDAO().upsert_diary(user_id, mood, body, formatted_date, diary_id)
+            formatted_date = datetime(year, month, int(day)).strftime('%Y-%m-%d')            
 
+            DiaryDAO().upsert_diary(user_id, mood, body, formatted_date, diary_id)
             # flash('일기가 성공적으로 저장되었습니다!', 'success')
             return redirect(url_for('diary.diary_home', year=year, month=month))
 
         elif _method == "delete":
-            user_id = 1
-
             DiaryDAO().delete_diary(diary_id)
             return redirect(url_for('diary.diary_home', year=year, month=month))
